remove lib module and add autogenerated graphql shit
