--- frontend/src/graphql/apollo-helpers.ts
@@ -1,238 +1,0 @@
-import { FieldPolicy, FieldReadFunction, TypePolicies, TypePolicy } from '@apollo/client/cache';
-export type QueryKeySpecifier = ('artist' | 'artistFromName' | 'collection' | 'collectionFromNameAndType' | 'playlist' | 'playlistFromNameAndType' | 'release' | 'track' | 'user' | 'artists' | 'collections' | 'playlists' | 'releases' | 'releaseYears' | QueryKeySpecifier)[];
-export type QueryFieldPolicy = {
-	artist?: FieldPolicy<any> | FieldReadFunction<any>,
-	artistFromName?: FieldPolicy<any> | FieldReadFunction<any>,
-	collection?: FieldPolicy<any> | FieldReadFunction<any>,
-	collectionFromNameAndType?: FieldPolicy<any> | FieldReadFunction<any>,
-	playlist?: FieldPolicy<any> | FieldReadFunction<any>,
-	playlistFromNameAndType?: FieldPolicy<any> | FieldReadFunction<any>,
-	release?: FieldPolicy<any> | FieldReadFunction<any>,
-	track?: FieldPolicy<any> | FieldReadFunction<any>,
-	user?: FieldPolicy<any> | FieldReadFunction<any>,
-	artists?: FieldPolicy<any> | FieldReadFunction<any>,
-	collections?: FieldPolicy<any> | FieldReadFunction<any>,
-	playlists?: FieldPolicy<any> | FieldReadFunction<any>,
-	releases?: FieldPolicy<any> | FieldReadFunction<any>,
-	releaseYears?: FieldPolicy<any> | FieldReadFunction<any>
-};
-export type MutationKeySpecifier = ('createArtist' | 'updateArtist' | 'createCollection' | 'updateCollection' | 'addReleaseToCollection' | 'delReleaseFromCollection' | 'createPlaylist' | 'updatePlaylist' | 'addTrackToPlaylist' | 'delTrackFromPlaylist' | 'createRelease' | 'updateRelease' | 'addArtistToRelease' | 'delArtistFromRelease' | 'updateTrack' | 'addArtistToTrack' | 'delArtistFromTrack' | 'updateUser' | 'newToken' | MutationKeySpecifier)[];
-export type MutationFieldPolicy = {
-	createArtist?: FieldPolicy<any> | FieldReadFunction<any>,
-	updateArtist?: FieldPolicy<any> | FieldReadFunction<any>,
-	createCollection?: FieldPolicy<any> | FieldReadFunction<any>,
-	updateCollection?: FieldPolicy<any> | FieldReadFunction<any>,
-	addReleaseToCollection?: FieldPolicy<any> | FieldReadFunction<any>,
-	delReleaseFromCollection?: FieldPolicy<any> | FieldReadFunction<any>,
-	createPlaylist?: FieldPolicy<any> | FieldReadFunction<any>,
-	updatePlaylist?: FieldPolicy<any> | FieldReadFunction<any>,
-	addTrackToPlaylist?: FieldPolicy<any> | FieldReadFunction<any>,
-	delTrackFromPlaylist?: FieldPolicy<any> | FieldReadFunction<any>,
-	createRelease?: FieldPolicy<any> | FieldReadFunction<any>,
-	updateRelease?: FieldPolicy<any> | FieldReadFunction<any>,
-	addArtistToRelease?: FieldPolicy<any> | FieldReadFunction<any>,
-	delArtistFromRelease?: FieldPolicy<any> | FieldReadFunction<any>,
-	updateTrack?: FieldPolicy<any> | FieldReadFunction<any>,
-	addArtistToTrack?: FieldPolicy<any> | FieldReadFunction<any>,
-	delArtistFromTrack?: FieldPolicy<any> | FieldReadFunction<any>,
-	updateUser?: FieldPolicy<any> | FieldReadFunction<any>,
-	newToken?: FieldPolicy<any> | FieldReadFunction<any>
-};
-export type ArtistKeySpecifier = ('id' | 'name' | 'starred' | 'numReleases' | 'imageId' | 'releases' | 'topGenres' | ArtistKeySpecifier)[];
-export type ArtistFieldPolicy = {
-	id?: FieldPolicy<any> | FieldReadFunction<any>,
-	name?: FieldPolicy<any> | FieldReadFunction<any>,
-	starred?: FieldPolicy<any> | FieldReadFunction<any>,
-	numReleases?: FieldPolicy<any> | FieldReadFunction<any>,
-	imageId?: FieldPolicy<any> | FieldReadFunction<any>,
-	releases?: FieldPolicy<any> | FieldReadFunction<any>,
-	topGenres?: FieldPolicy<any> | FieldReadFunction<any>
-};
-export type ArtistsKeySpecifier = ('results' | ArtistsKeySpecifier)[];
-export type ArtistsFieldPolicy = {
-	results?: FieldPolicy<any> | FieldReadFunction<any>
-};
-export type CollectionKeySpecifier = ('id' | 'name' | 'starred' | 'type' | 'numReleases' | 'lastUpdatedOn' | 'imageId' | 'releases' | 'topGenres' | CollectionKeySpecifier)[];
-export type CollectionFieldPolicy = {
-	id?: FieldPolicy<any> | FieldReadFunction<any>,
-	name?: FieldPolicy<any> | FieldReadFunction<any>,
-	starred?: FieldPolicy<any> | FieldReadFunction<any>,
-	type?: FieldPolicy<any> | FieldReadFunction<any>,
-	numReleases?: FieldPolicy<any> | FieldReadFunction<any>,
-	lastUpdatedOn?: FieldPolicy<any> | FieldReadFunction<any>,
-	imageId?: FieldPolicy<any> | FieldReadFunction<any>,
-	releases?: FieldPolicy<any> | FieldReadFunction<any>,
-	topGenres?: FieldPolicy<any> | FieldReadFunction<any>
-};
-export type CollectionsKeySpecifier = ('results' | CollectionsKeySpecifier)[];
-export type CollectionsFieldPolicy = {
-	results?: FieldPolicy<any> | FieldReadFunction<any>
-};
-export type PlaylistKeySpecifier = ('id' | 'name' | 'starred' | 'type' | 'numTracks' | 'lastUpdatedOn' | 'imageId' | 'tracks' | 'topGenres' | PlaylistKeySpecifier)[];
-export type PlaylistFieldPolicy = {
-	id?: FieldPolicy<any> | FieldReadFunction<any>,
-	name?: FieldPolicy<any> | FieldReadFunction<any>,
-	starred?: FieldPolicy<any> | FieldReadFunction<any>,
-	type?: FieldPolicy<any> | FieldReadFunction<any>,
-	numTracks?: FieldPolicy<any> | FieldReadFunction<any>,
-	lastUpdatedOn?: FieldPolicy<any> | FieldReadFunction<any>,
-	imageId?: FieldPolicy<any> | FieldReadFunction<any>,
-	tracks?: FieldPolicy<any> | FieldReadFunction<any>,
-	topGenres?: FieldPolicy<any> | FieldReadFunction<any>
-};
-export type PlaylistsKeySpecifier = ('results' | PlaylistsKeySpecifier)[];
-export type PlaylistsFieldPolicy = {
-	results?: FieldPolicy<any> | FieldReadFunction<any>
-};
-export type ReleaseKeySpecifier = ('id' | 'title' | 'releaseType' | 'addedOn' | 'inInbox' | 'inFavorites' | 'releaseYear' | 'releaseDate' | 'rating' | 'numTracks' | 'runtime' | 'imageId' | 'artists' | 'tracks' | 'genres' | 'labels' | 'collages' | ReleaseKeySpecifier)[];
-export type ReleaseFieldPolicy = {
-	id?: FieldPolicy<any> | FieldReadFunction<any>,
-	title?: FieldPolicy<any> | FieldReadFunction<any>,
-	releaseType?: FieldPolicy<any> | FieldReadFunction<any>,
-	addedOn?: FieldPolicy<any> | FieldReadFunction<any>,
-	inInbox?: FieldPolicy<any> | FieldReadFunction<any>,
-	inFavorites?: FieldPolicy<any> | FieldReadFunction<any>,
-	releaseYear?: FieldPolicy<any> | FieldReadFunction<any>,
-	releaseDate?: FieldPolicy<any> | FieldReadFunction<any>,
-	rating?: FieldPolicy<any> | FieldReadFunction<any>,
-	numTracks?: FieldPolicy<any> | FieldReadFunction<any>,
-	runtime?: FieldPolicy<any> | FieldReadFunction<any>,
-	imageId?: FieldPolicy<any> | FieldReadFunction<any>,
-	artists?: FieldPolicy<any> | FieldReadFunction<any>,
-	tracks?: FieldPolicy<any> | FieldReadFunction<any>,
-	genres?: FieldPolicy<any> | FieldReadFunction<any>,
-	labels?: FieldPolicy<any> | FieldReadFunction<any>,
-	collages?: FieldPolicy<any> | FieldReadFunction<any>
-};
-export type ReleasesKeySpecifier = ('total' | 'results' | ReleasesKeySpecifier)[];
-export type ReleasesFieldPolicy = {
-	total?: FieldPolicy<any> | FieldReadFunction<any>,
-	results?: FieldPolicy<any> | FieldReadFunction<any>
-};
-export type TrackKeySpecifier = ('id' | 'title' | 'duration' | 'trackNumber' | 'discNumber' | 'release' | 'artists' | TrackKeySpecifier)[];
-export type TrackFieldPolicy = {
-	id?: FieldPolicy<any> | FieldReadFunction<any>,
-	title?: FieldPolicy<any> | FieldReadFunction<any>,
-	duration?: FieldPolicy<any> | FieldReadFunction<any>,
-	trackNumber?: FieldPolicy<any> | FieldReadFunction<any>,
-	discNumber?: FieldPolicy<any> | FieldReadFunction<any>,
-	release?: FieldPolicy<any> | FieldReadFunction<any>,
-	artists?: FieldPolicy<any> | FieldReadFunction<any>
-};
-export type TrackArtistKeySpecifier = ('artist' | 'role' | TrackArtistKeySpecifier)[];
-export type TrackArtistFieldPolicy = {
-	artist?: FieldPolicy<any> | FieldReadFunction<any>,
-	role?: FieldPolicy<any> | FieldReadFunction<any>
-};
-export type TopGenreKeySpecifier = ('genre' | 'numMatches' | TopGenreKeySpecifier)[];
-export type TopGenreFieldPolicy = {
-	genre?: FieldPolicy<any> | FieldReadFunction<any>,
-	numMatches?: FieldPolicy<any> | FieldReadFunction<any>
-};
-export type UserKeySpecifier = ('id' | 'nickname' | UserKeySpecifier)[];
-export type UserFieldPolicy = {
-	id?: FieldPolicy<any> | FieldReadFunction<any>,
-	nickname?: FieldPolicy<any> | FieldReadFunction<any>
-};
-export type TokenKeySpecifier = ('hex' | TokenKeySpecifier)[];
-export type TokenFieldPolicy = {
-	hex?: FieldPolicy<any> | FieldReadFunction<any>
-};
-export type CollectionAndReleaseKeySpecifier = ('collection' | 'release' | CollectionAndReleaseKeySpecifier)[];
-export type CollectionAndReleaseFieldPolicy = {
-	collection?: FieldPolicy<any> | FieldReadFunction<any>,
-	release?: FieldPolicy<any> | FieldReadFunction<any>
-};
-export type PlaylistAndTrackKeySpecifier = ('playlist' | 'track' | PlaylistAndTrackKeySpecifier)[];
-export type PlaylistAndTrackFieldPolicy = {
-	playlist?: FieldPolicy<any> | FieldReadFunction<any>,
-	track?: FieldPolicy<any> | FieldReadFunction<any>
-};
-export type ReleaseAndArtistKeySpecifier = ('release' | 'artist' | ReleaseAndArtistKeySpecifier)[];
-export type ReleaseAndArtistFieldPolicy = {
-	release?: FieldPolicy<any> | FieldReadFunction<any>,
-	artist?: FieldPolicy<any> | FieldReadFunction<any>
-};
-export type TrackAndArtistKeySpecifier = ('track' | 'trackArtist' | TrackAndArtistKeySpecifier)[];
-export type TrackAndArtistFieldPolicy = {
-	track?: FieldPolicy<any> | FieldReadFunction<any>,
-	trackArtist?: FieldPolicy<any> | FieldReadFunction<any>
-};
-export type TypedTypePolicies = TypePolicies & {
-	Query?: Omit<TypePolicy, "fields" | "keyFields"> & {
-		keyFields?: false | QueryKeySpecifier | (() => undefined | QueryKeySpecifier),
-		fields?: QueryFieldPolicy,
-	},
-	Mutation?: Omit<TypePolicy, "fields" | "keyFields"> & {
-		keyFields?: false | MutationKeySpecifier | (() => undefined | MutationKeySpecifier),
-		fields?: MutationFieldPolicy,
-	},
-	Artist?: Omit<TypePolicy, "fields" | "keyFields"> & {
-		keyFields?: false | ArtistKeySpecifier | (() => undefined | ArtistKeySpecifier),
-		fields?: ArtistFieldPolicy,
-	},
-	Artists?: Omit<TypePolicy, "fields" | "keyFields"> & {
-		keyFields?: false | ArtistsKeySpecifier | (() => undefined | ArtistsKeySpecifier),
-		fields?: ArtistsFieldPolicy,
-	},
-	Collection?: Omit<TypePolicy, "fields" | "keyFields"> & {
-		keyFields?: false | CollectionKeySpecifier | (() => undefined | CollectionKeySpecifier),
-		fields?: CollectionFieldPolicy,
-	},
-	Collections?: Omit<TypePolicy, "fields" | "keyFields"> & {
-		keyFields?: false | CollectionsKeySpecifier | (() => undefined | CollectionsKeySpecifier),
-		fields?: CollectionsFieldPolicy,
-	},
-	Playlist?: Omit<TypePolicy, "fields" | "keyFields"> & {
-		keyFields?: false | PlaylistKeySpecifier | (() => undefined | PlaylistKeySpecifier),
-		fields?: PlaylistFieldPolicy,
-	},
-	Playlists?: Omit<TypePolicy, "fields" | "keyFields"> & {
-		keyFields?: false | PlaylistsKeySpecifier | (() => undefined | PlaylistsKeySpecifier),
-		fields?: PlaylistsFieldPolicy,
-	},
-	Release?: Omit<TypePolicy, "fields" | "keyFields"> & {
-		keyFields?: false | ReleaseKeySpecifier | (() => undefined | ReleaseKeySpecifier),
-		fields?: ReleaseFieldPolicy,
-	},
-	Releases?: Omit<TypePolicy, "fields" | "keyFields"> & {
-		keyFields?: false | ReleasesKeySpecifier | (() => undefined | ReleasesKeySpecifier),
-		fields?: ReleasesFieldPolicy,
-	},
-	Track?: Omit<TypePolicy, "fields" | "keyFields"> & {
-		keyFields?: false | TrackKeySpecifier | (() => undefined | TrackKeySpecifier),
-		fields?: TrackFieldPolicy,
-	},
-	TrackArtist?: Omit<TypePolicy, "fields" | "keyFields"> & {
-		keyFields?: false | TrackArtistKeySpecifier | (() => undefined | TrackArtistKeySpecifier),
-		fields?: TrackArtistFieldPolicy,
-	},
-	TopGenre?: Omit<TypePolicy, "fields" | "keyFields"> & {
-		keyFields?: false | TopGenreKeySpecifier | (() => undefined | TopGenreKeySpecifier),
-		fields?: TopGenreFieldPolicy,
-	},
-	User?: Omit<TypePolicy, "fields" | "keyFields"> & {
-		keyFields?: false | UserKeySpecifier | (() => undefined | UserKeySpecifier),
-		fields?: UserFieldPolicy,
-	},
-	Token?: Omit<TypePolicy, "fields" | "keyFields"> & {
-		keyFields?: false | TokenKeySpecifier | (() => undefined | TokenKeySpecifier),
-		fields?: TokenFieldPolicy,
-	},
-	CollectionAndRelease?: Omit<TypePolicy, "fields" | "keyFields"> & {
-		keyFields?: false | CollectionAndReleaseKeySpecifier | (() => undefined | CollectionAndReleaseKeySpecifier),
-		fields?: CollectionAndReleaseFieldPolicy,
-	},
-	PlaylistAndTrack?: Omit<TypePolicy, "fields" | "keyFields"> & {
-		keyFields?: false | PlaylistAndTrackKeySpecifier | (() => undefined | PlaylistAndTrackKeySpecifier),
-		fields?: PlaylistAndTrackFieldPolicy,
-	},
-	ReleaseAndArtist?: Omit<TypePolicy, "fields" | "keyFields"> & {
-		keyFields?: false | ReleaseAndArtistKeySpecifier | (() => undefined | ReleaseAndArtistKeySpecifier),
-		fields?: ReleaseAndArtistFieldPolicy,
-	},
-	TrackAndArtist?: Omit<TypePolicy, "fields" | "keyFields"> & {
-		keyFields?: false | TrackAndArtistKeySpecifier | (() => undefined | TrackAndArtistKeySpecifier),
-		fields?: TrackAndArtistFieldPolicy,
-	}
-};
