+++ frontend/src/graphql/artists.graphql
@@ -1,0 +1,19 @@
+query FetchArtists {
+  artists {
+    results {
+      ...ArtistFields
+    }
+  }
+}
+
+query FetchArtist($id: Int!) {
+  artist(id: $id) {
+    ...ArtistFields
+  }
+}
+
+mutation UpdateArtist($id: Int!, $name: String, $starred: Boolean) {
+  updateArtist(id: $id, name: $name, starred: $starred) {
+    ...ArtistFields
+  }
+}
