+++ frontend/src/graphql/users.graphql
@@ -1,0 +1,11 @@
+query FetchUser {
+  user {
+    ...UserFields
+  }
+}
+
+mutation UpdateUser($nickname: String) {
+  updateUser(nickname: $nickname) {
+    ...UserFields
+  }
+}
