--- frontend/src/lib/artists/fetchAll.ts
@@ -1,20 +1,0 @@
-import { gql, QueryHookOptions, QueryResult, useQuery } from '@apollo/client';
-
-import { ARTIST_FIELDS } from '~/lib/fragments';
-import { ArtistT } from '~/types';
-
-const QUERY = gql`
-  query {
-    artists {
-      results {
-        ...ArtistFields
-      }
-    }
-  }
-  ${ARTIST_FIELDS}
-`;
-
-type T = { artists: { results: ArtistT[] } };
-
-export const useFetchArtists = (options?: QueryHookOptions<T>): QueryResult<T> =>
-  useQuery<T>(QUERY, options);
