--- frontend/src/lib/artists/index.ts
@@ -1,3 +1,0 @@
-export * from './fetchAll';
-export * from './fetchOne';
-export * from './mutateOne';
