--- frontend/src/lib/collections/index.ts
@@ -1,5 +1,0 @@
-export * from './addRelease';
-export * from './delRelease';
-export * from './fetchMany';
-export * from './fetchOne';
-export * from './mutateOne';
