--- frontend/src/lib/index.ts
@@ -1,4 +1,0 @@
-export * from './artists';
-export * from './collections';
-export * from './releases';
-export * from './user';
