--- frontend/src/lib/releases/index.ts
@@ -1,5 +1,0 @@
-export * from './fetchOne';
-export * from './fetchYears';
-export * from './mutateOne';
-export * from './recentlyAdded';
-export * from './search';
