--- frontend/src/lib/user/index.ts
@@ -1,2 +1,0 @@
-export * from './fetch';
-export * from './mutate';
